fix(_rmsh): C api fixes
Firstly, now initial guess for all positions is zeros. This prevents errors that would occur before when memory had Nans, which would upset the solver.

Second, most local variables that were dealing with counts and offsets are now size_t instead of unsigned. This is probably not needed, but just in case.

Lastly, since module is no longer being compiled to be ABI compatible, the type is defined using a PyTypeObject. This makes initialization simpler.